+++ app/src/main/java/com/orgzly/android/ui/tags/TagBrowserRow.kt
@@ -1,0 +1,6 @@
+package com.orgzly.android.ui.tags
+
+data class TagBrowserRow(
+        val tag: String,
+        val activeCount: Int
+)